fix common errors

+++ first/src/main/java/com/pavel/alex/lab/first/generator/SecureJavaGenerator.java
@@ -1,6 +1,6 @@
 package com.pavel.alex.lab.first.generator;
 
-import java.util.Random;
+import java.security.SecureRandom;
 /*
 A cryptographically strong random number minimally complies with the statistical random number generator tests specified in FIPS 140-2, Security Requirements for Cryptographic Modules, section 4.9.1. Additionally, SecureRandom must produce non-deterministic output. Therefore any seed material passed to a SecureRandom object must be unpredictable, and all SecureRandom output sequences must be cryptographically strong, as described in RFC 1750: Randomness Recommendations for Security.
 */
@@ -8,12 +8,12 @@ public class SecureJavaGenerator implements Generator {
 
     private SecureRandom random;
     private byte bytes[];
-    private long current = 0;
+    private int current = 0;
 
     public SecureJavaGenerator(int count) {
         random = new SecureRandom();
         bytes = new byte[count];
-        byte seed[] = random.generateSeed(System.currentTimeMillis());
+        byte seed[] = random.generateSeed((int) System.currentTimeMillis());
         random.nextBytes(bytes);
     }
